requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
ipdb>=0.12.2
mock>=3.0.5
PyJWT>=1.7.1
pytest-django==3.5.1
factory-boy>=2.12.0
coverage<5.0
pytest-cov>=2.7.1
coveralls==1.8.2
boto3>=1.9
jsonschema>=3.0.0
django-cors-headers>=3.2.0
numpy-stl==2.10.1
geopy>=1.21.0
mercadopago>=1.1.1